requirements.txt
tavily-python
duckduckgo-search
datasets
python-dotenv
pydantic
